setting decoder layer for parseq
