+++ doctr/models/utils/pytorch.py
@@ -11,21 +11,8 @@ from torch import nn
 
 from doctr.utils.data import download_from_url
 
-__all__ = ["load_pretrained_params_local","load_pretrained_params", "conv_sequence_pt", "export_model_to_onnx"]
+__all__ = ["load_pretrained_params", "conv_sequence_pt", "export_model_to_onnx"]
 
-"""
-def load_pretrained_params_local(model,path):
-    state_dict = torch.load(path, map_location="cpu")
-    state_dict_corr =  {}
-
-    for key in state_dict:
-        state_dict_corr['feat_extractor.'+key]=state_dict[key]
-        state_dict_corr[key]=state_dict[key]
-    model.load_state_dict(state_dict_corr, strict=False)
-"""
-def load_pretrained_params_local(model,path):
-    state_dict = torch.load(path, map_location="cpu")
-    model.load_state_dict(state_dict)
     
 def load_pretrained_params(
     model: nn.Module,
